// frontend/src/client/types.gen.ts
export type AppointmentCreate = {
    patient_name: string;
    patient_email: string;
    reason_for_visit?: (string | null);
    doctor_id: string;
    availability_id: string;
};
// ...
export type DoctorAvailabilityPublic = {
    start_time: string;
    end_time: string;
    id: string;
    doctor_id: string;
    is_booked: boolean;
};

export type DoctorCreate = {
    full_name: string;
    specialty: string;
    bio?: (string | null);
};

export type DoctorPublic = {
    full_name: string;
    specialty: string;
    bio?: (string | null);
    id: string;
};